Add Cause to events
