+++ FlightControl/src/main/java/org/spazzinq/flightcontrol/check/always/AdvancedEnchantmentsCheck.java
@@ -28,7 +28,7 @@ import n3kas.ae.api.AEAPI;
 import org.bukkit.entity.Player;
 import org.spazzinq.flightcontrol.FlightControl;
 import org.spazzinq.flightcontrol.check.Check;
-import org.spazzinq.flightcontrol.object.Cause;
+import org.spazzinq.flightcontrol.api.object.Cause;
 
 public class AdvancedEnchantmentsCheck extends Check {
     private FlightControl pl;
