+++ FlightControl/src/main/java/org/spazzinq/flightcontrol/check/bypasstrail/InvisibilityPotionCheck.java
@@ -27,7 +27,7 @@ package org.spazzinq.flightcontrol.check.bypasstrail;
 import org.bukkit.entity.Player;
 import org.bukkit.potion.PotionEffectType;
 import org.spazzinq.flightcontrol.check.Check;
-import org.spazzinq.flightcontrol.object.Cause;
+import org.spazzinq.flightcontrol.api.object.Cause;
 
 public class InvisibilityPotionCheck extends Check {
     @Override public boolean check(Player p) {
