+++ FlightControl/src/main/java/org/spazzinq/flightcontrol/check/category/CategoryCheck.java
@@ -25,7 +25,7 @@
 package org.spazzinq.flightcontrol.check.category;
 
 import org.spazzinq.flightcontrol.check.Check;
-import org.spazzinq.flightcontrol.object.Cause;
+import org.spazzinq.flightcontrol.api.object.Cause;
 
 public abstract class CategoryCheck extends Check {
     @Override public Cause getCause() {
