+++ FlightControl/src/main/java/org/spazzinq/flightcontrol/check/combat/CombatCheck.java
@@ -25,7 +25,7 @@
 package org.spazzinq.flightcontrol.check.combat;
 
 import org.spazzinq.flightcontrol.check.Check;
-import org.spazzinq.flightcontrol.object.Cause;
+import org.spazzinq.flightcontrol.api.object.Cause;
 
 public abstract class CombatCheck extends Check {
     @Override public Cause getCause() {
