+++ FlightControl/src/main/java/org/spazzinq/flightcontrol/check/territory/TerritoryCheck.java
@@ -25,7 +25,7 @@
 package org.spazzinq.flightcontrol.check.territory;
 
 import org.spazzinq.flightcontrol.check.Check;
-import org.spazzinq.flightcontrol.object.Cause;
+import org.spazzinq.flightcontrol.api.object.Cause;
 
 public abstract class TerritoryCheck extends Check {
     @Override public Cause getCause() {
